Add files via upload
This notebook leverage the model created in step 4 to project Blue Mountain's ticket pricing and impact on pricing of changes in various features.It also provides a summary on the model to the business and how best to interpret the results. Finally, it provides recommendations for the model's use and further work that would strengthen the model.